Fix: set assertions to not use local variables to remove warnings in Release build
Previously we assigned local variables before assertions and some of these
local variables were only used in assertions. Because of that, when we build in
Release mode and the assertions are not called, we get a warning that a
variable is declared but never used. Now there's no such warning